Backend/core-backend/requirements.txt
torch==2.7.0
torchvision
numpy
pillow
scikit-learn
matplotlib
pandas
Flask
Flask-Cors
gunicorn
tqdm
requests
pymongo==4.6.3
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2 